+++ zcash_client_sqlite/src/testing.rs
@@ -14,20 +14,21 @@ use tempfile::NamedTempFile;
 #[cfg(feature = "unstable")]
 use tempfile::TempDir;
 
-use zcash_client_backend::data_api::chain::ScanSummary;
-use zcash_client_backend::data_api::{AccountBalance, WalletRead};
+use zcash_client_backend::fees::{standard, DustOutputPolicy};
 #[allow(deprecated)]
 use zcash_client_backend::{
     address::RecipientAddress,
     data_api::{
         self,
-        chain::{scan_cached_blocks, BlockSource},
+        chain::{scan_cached_blocks, BlockSource, ScanSummary},
         wallet::{
             create_proposed_transaction, create_spend_to_address,
-            input_selection::{GreedyInputSelectorError, InputSelector, Proposal},
-            propose_transfer, spend,
+            input_selection::{
+                GreedyInputSelector, GreedyInputSelectorError, InputSelector, Proposal,
+            },
+            propose_standard_transfer_to_address, propose_transfer, spend,
         },
-        AccountBirthday, WalletSummary, WalletWrite,
+        AccountBalance, AccountBirthday, WalletRead, WalletSummary, WalletWrite,
     },
     keys::UnifiedSpendingKey,
     proto::compact_formats::{
@@ -40,7 +41,7 @@ use zcash_note_encryption::Domain;
 use zcash_primitives::{
     block::BlockHash,
     consensus::{self, BlockHeight, Network, NetworkUpgrade, Parameters},
-    memo::MemoBytes,
+    memo::{Memo, MemoBytes},
     sapling::{
         note_encryption::{sapling_note_encryption, SaplingDomain},
         util::generate_random_rseed,
@@ -48,8 +49,8 @@ use zcash_primitives::{
         Note, Nullifier, PaymentAddress,
     },
     transaction::{
-        components::amount::{BalanceError, NonNegativeAmount},
-        fees::FeeRule,
+        components::amount::NonNegativeAmount,
+        fees::{zip317::FeeError as Zip317FeeError, FeeRule, StandardFeeRule},
         Transaction, TxId,
     },
     zip32::{sapling::DiversifiableFullViewingKey, DiversifierIndex},
@@ -442,8 +443,8 @@ impl<Cache> TestState<Cache> {
         data_api::error::Error<
             SqliteClientError,
             commitment_tree::Error,
-            GreedyInputSelectorError<BalanceError, ReceivedNoteId>,
-            Infallible,
+            GreedyInputSelectorError<Zip317FeeError, ReceivedNoteId>,
+            Zip317FeeError,
         >,
     > {
         let params = self.network();
@@ -526,6 +527,41 @@ impl<Cache> TestState<Cache> {
         )
     }
 
+    /// Invokes [`propose_standard_transfer`] with the given arguments.
+    #[allow(clippy::type_complexity)]
+    #[allow(clippy::too_many_arguments)]
+    pub(crate) fn propose_standard_transfer<CommitmentTreeErrT>(
+        &mut self,
+        spend_from_account: AccountId,
+        fee_rule: StandardFeeRule,
+        min_confirmations: NonZeroU32,
+        to: &RecipientAddress,
+        amount: NonNegativeAmount,
+        memo: Option<MemoBytes>,
+        change_memo: Option<MemoBytes>,
+    ) -> Result<
+        Proposal<StandardFeeRule, ReceivedNoteId>,
+        data_api::error::Error<
+            SqliteClientError,
+            CommitmentTreeErrT,
+            GreedyInputSelectorError<Zip317FeeError, ReceivedNoteId>,
+            Zip317FeeError,
+        >,
+    > {
+        let params = self.network();
+        propose_standard_transfer_to_address::<_, _, CommitmentTreeErrT>(
+            &mut self.db_data,
+            &params,
+            fee_rule,
+            spend_from_account,
+            min_confirmations,
+            to,
+            amount,
+            memo,
+            change_memo,
+        )
+    }
+
     /// Invokes [`propose_shielding`] with the given arguments.
     #[cfg(feature = "transparent-inputs")]
     #[allow(clippy::type_complexity)]
@@ -560,7 +596,7 @@ impl<Cache> TestState<Cache> {
     }
 
     /// Invokes [`create_proposed_transaction`] with the given arguments.
-    pub(crate) fn create_proposed_transaction<FeeRuleT>(
+    pub(crate) fn create_proposed_transaction<InputsErrT, FeeRuleT>(
         &mut self,
         usk: &UnifiedSpendingKey,
         ovk_policy: OvkPolicy,
@@ -571,7 +607,7 @@ impl<Cache> TestState<Cache> {
         data_api::error::Error<
             SqliteClientError,
             commitment_tree::Error,
-            Infallible,
+            InputsErrT,
             FeeRuleT::Error,
         >,
     >
@@ -579,7 +615,7 @@ impl<Cache> TestState<Cache> {
         FeeRuleT: FeeRule,
     {
         let params = self.network();
-        create_proposed_transaction::<_, _, Infallible, _>(
+        create_proposed_transaction(
             &mut self.db_data,
             &params,
             test_prover(),
@@ -994,4 +1030,16 @@ impl TestCache for FsBlockCache {
 
         meta
     }
+}
+
+pub(crate) fn input_selector(
+    fee_rule: StandardFeeRule,
+    change_memo: Option<&str>,
+) -> GreedyInputSelector<
+    WalletDb<rusqlite::Connection, Network>,
+    standard::SingleOutputChangeStrategy,
+> {
+    let change_memo = change_memo.map(|m| MemoBytes::from(m.parse::<Memo>().unwrap()));
+    let change_strategy = standard::SingleOutputChangeStrategy::new(fee_rule, change_memo);
+    GreedyInputSelector::new(change_strategy, DustOutputPolicy::default())
 }